clean module
