+++ .idea/gradle.xml
@@ -10,10 +10,8 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/benchmark" />
-            <option value="$PROJECT_DIR$/bootstrap" />
             <option value="$PROJECT_DIR$/core" />
             <option value="$PROJECT_DIR$/monitor" />
-            <option value="$PROJECT_DIR$/recovery" />
             <option value="$PROJECT_DIR$/titan-stomp" />
           </set>
         </option>
