--- bootstrap/build.gradle.kts
@@ -1,6 +1,0 @@
-group = "org.traffichunter.titan.bootstrap"
-version = "1.0-SNAPSHOT"
-
-dependencies {
-    implementation(project.libs.snakeyaml)
-}
