--- bootstrap/src/main/java/org/traffichunter/titan/bootstrap/BootState.java
@@ -1,45 +1,0 @@
-/*
- * The MIT License
- *
- * Copyright (c) 2024 traffic-hunter
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-package org.traffichunter.titan.bootstrap;
-
-import java.util.concurrent.atomic.AtomicBoolean;
-
-/**
- * @author yungwang-o
- */
-public final class BootState {
-
-    private static final Boolean STATE_NONE = false;
-    private static final Boolean STATE_STARTED = true;
-
-    private final AtomicBoolean state = new AtomicBoolean(STATE_NONE);
-
-    boolean getState() {
-        return state.get();
-    }
-
-    public boolean start() {
-        return state.compareAndSet(STATE_NONE, STATE_STARTED);
-    }
-}
