--- bootstrap/src/main/java/org/traffichunter/titan/bootstrap/environment/RelaxedBindingUtils.java
@@ -1,55 +1,0 @@
-/**
- * The MIT License
- *
- * Copyright (c) 2024 yungwang-o
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-package org.traffichunter.titan.bootstrap.environment;
-
-import org.yaml.snakeyaml.introspector.Property;
-import org.yaml.snakeyaml.introspector.PropertyUtils;
-
-/**
- * @author yungwang-o
- * @version 1.0.0
- */
-final class RelaxedBindingUtils extends PropertyUtils {
-
-    @Override
-    public Property getProperty(final Class<?> type, final String name) {
-        return super.getProperty(type, kebabToCamel(name));
-    }
-
-    private String kebabToCamel(final String kebab) {
-        final StringBuilder sb = new StringBuilder();
-
-        String[] split = kebab.split("-");
-
-        for(int i = 0; i < split.length; i++) {
-            if(i == 0) {
-                sb.append(split[i]);
-                continue;
-            }
-            sb.append(split[i].replaceFirst("^[a-z]", String.valueOf(split[i].charAt(0)).toUpperCase()));
-        }
-
-        return sb.toString();
-    }
-}
