+++ monitor/build.gradle.kts
@@ -2,5 +2,4 @@ group = "org.traffichunter.titan.monitor"
 version = "1.0-SNAPSHOT"
 
 dependencies {
-    implementation(project(":bootstrap"))
 }
