+++ monitor/src/main/java/org/traffichunter/titan/monitor/Monitor.java
@@ -23,137 +23,11 @@
  */
 package org.traffichunter.titan.monitor;
 
-import java.util.Map;
-import java.util.concurrent.Executors;
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.ThreadFactory;
-import java.util.concurrent.TimeUnit;
 import lombok.extern.slf4j.Slf4j;
-import org.traffichunter.titan.bootstrap.event.EventBusHolder;
-import org.traffichunter.titan.bootstrap.monitor.SettingsMonitor;
-import org.traffichunter.titan.monitor.jmx.JmxMbeanCollector;
-import org.traffichunter.titan.monitor.jmx.cpu.CpuData;
-import org.traffichunter.titan.monitor.jmx.cpu.JmxCpuMbeanCollector;
-import org.traffichunter.titan.monitor.jmx.heap.HeapData;
-import org.traffichunter.titan.monitor.jmx.heap.JmxHeapMbeanCollector;
-import org.traffichunter.titan.monitor.jmx.thread.JmxThreadMbeanCollector;
-import org.traffichunter.titan.monitor.jmx.thread.ThreadData;
 
 /**
  * @author yungwang-o
  */
 @Slf4j
 public final class Monitor {
-
-    private static final long DEFAULT_SHUTDOWN_TIMEOUT = 10;
-
-    private final Map<Class<?>, JmxMbeanCollector<?>> collectors;
-
-    private final ScheduledExecutorService schedule;
-
-    private final EventBusHolder eventBusHolder = EventBusHolder.INSTANCE;
-
-    private final SettingsMonitor settingsMonitor;
-
-    public Monitor(final SettingsMonitor settingsMonitor) {
-
-        if(settingsMonitor.scheduledThreadPool() == 0) {
-            throw new IllegalArgumentException("Scheduled thread pool must be non-zero");
-        }
-
-        if(settingsMonitor.scheduledThreadPool() == 1) {
-            this.schedule = Executors.newSingleThreadScheduledExecutor(createMonitorThreadFactory());
-        } else {
-            this.schedule = Executors.newScheduledThreadPool(
-                    settingsMonitor.scheduledThreadPool(),
-                    createMonitorThreadFactory()
-            );
-        }
-
-        this.settingsMonitor = settingsMonitor;
-        this.collectors = Map.of(
-                CpuData.class, new JmxCpuMbeanCollector(),
-                HeapData.class, new JmxHeapMbeanCollector(),
-                ThreadData.class, new JmxThreadMbeanCollector()
-        );
-    }
-
-    public <R> void doMonitor(final Class<R> dataType) {
-
-        JmxMbeanCollector<R> collector = findMonitor(dataType);
-
-        schedule.scheduleWithFixedDelay(() -> {
-
-                    R collect = collector.collect();
-
-                    if(isAppHealthThresholdExceeded(collect)) {
-                        eventBusHolder.post(collect);
-                    }
-                },
-                settingsMonitor.initialDelay(),
-                settingsMonitor.delay(),
-                settingsMonitor.timeUnit()
-        );
-    }
-
-    public void close() {
-
-        schedule.shutdown();
-
-        try {
-            if (!schedule.awaitTermination(DEFAULT_SHUTDOWN_TIMEOUT, TimeUnit.SECONDS)) {
-                schedule.shutdownNow();
-            }
-            log.info("Monitor shutdown completed");
-        } catch (InterruptedException e) {
-            schedule.shutdownNow();
-            Thread.currentThread().interrupt();
-            log.error("Monitor shutdown interrupted", e);
-        }
-    }
-
-    @SuppressWarnings("unchecked")
-    public <R> JmxMbeanCollector<R> findMonitor(final Class<R> dataType) {
-        JmxMbeanCollector<?> collector = collectors.get(dataType);
-
-        if(collector == null || !dataType.equals(collector.getDataType())) {
-            throw new IllegalStateException("Not found collector for type " + dataType.getSimpleName());
-        }
-
-        return (JmxMbeanCollector<R>) collector;
-    }
-
-    private <R> boolean isAppHealthThresholdExceeded(final R data) {
-
-        switch (data) {
-            case CpuData cpuData -> {
-                return cpuData.isCheckThreshold(80);
-            }
-            case HeapData heapData -> {
-                return heapData.isCheckThreshold(0.8);
-            }
-            case ThreadData threadData -> {
-                return threadData.isCheckThreshold(100);
-            }
-            case null, default -> {
-                return false;
-            }
-        }
-    }
-
-    private ThreadFactory createMonitorThreadFactory() {
-        return run -> {
-            Thread thread = new Thread(run, "MonitorThread");
-            thread.setDaemon(true);
-            thread.setUncaughtExceptionHandler(
-                    (thread1, throwable) ->
-                            log.info(
-                                    "[{}] Monitor thread exception = {}",
-                                    thread1.getName(),
-                                    throwable.getMessage()
-                            )
-            );
-            return thread;
-        };
-    }
 }
