--- recovery/build.gradle.kts
@@ -1,13 +1,0 @@
-plugins {
-    id("java")
-}
-
-group = "org.traffichunter.titan.recovery"
-version = "1.0-SNAPSHOT"
-
-dependencies {
-}
-
-tasks.test {
-    useJUnitPlatform()
-}
