+++ settings.gradle.kts
@@ -1,9 +1,7 @@
 rootProject.name = "titan"
 
-include("bootstrap")
 include("core")
 include("monitor")
-include("recovery")
 include("titan-stomp")
 
 include("benchmark")
